Change response from annotations to classification
Previous ELG version contained extra features that are not in the
original command-line tool, and the usage was limited to the top
language. Now n-best languages is returned and hopefully this gives
more options to the potential user.

HeLI.java code base is updated to version 1.4. In addition some
code related to command-line handling is removed.